requirements.txt
python-dotenv
setuptools
scikit-learn
pandas
numpy
matplotlib
seaborn
logging
exception
pymongo[srv]==3.12
pymongo
certifi
pyyaml
xgboost
mlflow
dagshub
joblib
fastapi
uvicorn
python-multipart

# -e . 
